+++ app/app.go
@@ -2,5 +2,5 @@ package app
 
 const (
 	Name    = "circumflex"
-	Version = "2.6-dev"
+	Version = "2.6"
 )